requirements.txt
nest-asyncio==1.6.0
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.32.0
streamlit-autorefresh==1.0.1
telethon==1.34.0
